fix: No tests are found when no build.gradle[.kts] file present in root
When there's no build.gradle[.kts] file present in a root folder of the
project (which is the case if `gradle init` command is used),
neotest-gradle can't run tests in this case.

Another issue is that if an empty build.gradle[.kts] file is added to
the root, neotest-gradle still can't run tests because the filter
command turns out to be invalid.

This PR is fixing both issues.